+++ tests/C++VariantTest.cpp
@@ -9,6 +9,7 @@ namespace CppVariantTests
 //####### Enum Test #######
 TEST(EnumTest, testEnumValues)
 {
+	ASSERT_EQ(HttpStatus::Invalid,             -1);
 	ASSERT_EQ(HttpStatus::OK,                  200);
 	ASSERT_EQ(HttpStatus::NotFound,            404);
 	ASSERT_EQ(HttpStatus::InternalServerError, 500);
@@ -18,6 +19,7 @@ TEST(EnumTest, testEnumValues)
 //####### Reason Phrase Test #######
 TEST(ReasonPhraseTest, testEnumParameter)
 {
+	ASSERT_EQ(HttpStatus::reasonPhrase(HttpStatus::Invalid),             std::string());
 	ASSERT_EQ(HttpStatus::reasonPhrase(HttpStatus::OK),                  std::string("OK"));
 	ASSERT_EQ(HttpStatus::reasonPhrase(HttpStatus::NotFound),            std::string("Not Found"));
 	ASSERT_EQ(HttpStatus::reasonPhrase(HttpStatus::InternalServerError), std::string("Internal Server Error"));
@@ -25,6 +27,7 @@ TEST(ReasonPhraseTest, testEnumParameter)
 
 TEST(ReasonPhraseTest, testIntegerParameter)
 {
+	ASSERT_EQ(HttpStatus::reasonPhrase(static_cast<int>(HttpStatus::Invalid)),        std::string());
 	ASSERT_EQ(HttpStatus::reasonPhrase(static_cast<int>(HttpStatus::Created)),        std::string("Created"));
 	ASSERT_EQ(HttpStatus::reasonPhrase(static_cast<int>(HttpStatus::Unauthorized)),   std::string("Unauthorized"));
 	ASSERT_EQ(HttpStatus::reasonPhrase(static_cast<int>(HttpStatus::GatewayTimeout)), std::string("Gateway Timeout"));
@@ -34,6 +37,7 @@ TEST(ReasonPhraseTest, testIntegerParameter)
 //####### Category Tester Test #######
 INSTANTIATE_TEST_CASE_P(DefaultInstance, CategoryTesterTest, ::testing::Values(
 //                   // code                          // info // success // redir // clientErr // serverErr // error
+CategoryTesterParams(HttpStatus::Invalid,             false,  false,     false,   false,       false,       false),
 CategoryTesterParams(HttpStatus::SwitchingProtocols,  true,   false,     false,   false,       false,       false),
 CategoryTesterParams(HttpStatus::NoContent,           false,  true,      false,   false,       false,       false),
 CategoryTesterParams(HttpStatus::SeeOther,            false,  false,     true,    false,       false,       false),