+++ CHANGELOG.md
@@ -13,7 +13,7 @@ This changelog follows the [Keep a Changelog](http://keepachangelog.com) format.
 
 ### Added ###
 - [#28] Added code `Invalid` (-1).
-- [#28]{Qt C++11} Qt C++11 variant with uses `enum class` for the codes.
+- [#28]{Qt C++11} Qt C++11 variant which uses `enum class` for the codes.
 
 ### Changed ###
 - {Qt} `networkErrorToStatusCode()` now returns a `Code` instead of `int`.