+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This changelog follows the [Keep a Changelog](http://keepachangelog.com) format.
 
 ---
 
-## [Unreleased]
+## [1.6.0] - 2025-02-09
 
 ### Added ###
 - [#28] Added code `Invalid` (-1).
@@ -106,9 +106,12 @@ Initial (actually unversioned) release.
 ---
 
 
+[1.6.0]: https://github.com/j-ulrich/http-status-codes-cpp/compare/1.5.0...1.6.0
 [1.5.0]: https://github.com/j-ulrich/http-status-codes-cpp/releases/tag/1.5.0
 [1.4.0]: https://github.com/j-ulrich/http-status-codes-cpp/releases/tag/1.4.0
 [1.3.0]: https://github.com/j-ulrich/http-status-codes-cpp/releases/tag/1.3.0
 [1.2.0]: https://github.com/j-ulrich/http-status-codes-cpp/releases/tag/1.2.0
 [1.1.1]: https://github.com/j-ulrich/http-status-codes-cpp/releases/tag/1.1.1
 [1.1.0]: https://github.com/j-ulrich/http-status-codes-cpp/releases/tag/1.1.0
+
+[unreleased]: https://github.com/j-ulrich/http-status-codes-cpp/compare/1.6.0...develop